Implemented PCA to reduce feature image dimensionality, and made a clustering algorithm and cluster search algorithm to make the search go faster. The symbol set is small so I shouldn't get excited, but compared to before it is a couple orders of magnitude more efficient. So I must be doing something right because that is the speedup the original paper cited.
Next things to do: I need to store the results of the PCA calculation and probably also the cluster-sorting, since those are several-second-long operations that should not change from run to run, unless the user is adding tons of new symbols. I also have to think about implementing rotation (in)variance, and possibly include a flag for symbols indicating 'this way up', which is important for handwriting where orientation has semantics.